new version of stride - switch based
Looking at asm it became obvious to me that this benchmark is no good
It's ALL about compiler able to inline or not certain things,
e.g. only optStride and stride are properly inlined.
Results depend on phase of moon and the mood of optimizer
I gave up untill GDC compiles uni.d.